rename StudioPlugin and QueryPlugin
