+++ .changeset/stale-dolls-dance.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-studio': patch
+---
+
+**BREAKING CHANGE:** Change `StudioPlugin` to `LegendStudioPlugin`.
