+++ .changeset/tall-moons-do.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-query': patch
+---
+
+**BREAKING CHANGE:** Change `QueryPlugin` to `LegendQueryPlugin`.
