+++ packages/legend-extension-dsl-data-space/src/components/query/DSLDataSpace_QueryPlugin.tsx
@@ -22,12 +22,12 @@ import type {
   QuerySetupState,
   QuerySetupStore,
 } from '@finos/legend-query';
-import { QueryPlugin } from '@finos/legend-query';
+import { LegendQueryPlugin } from '@finos/legend-query';
 import { SquareIcon } from '@finos/legend-art';
 import { DataSpaceQuerySetupState } from '../../stores/query/DataSpaceQuerySetupState';
 import { DataspaceQuerySetup } from './DataSpaceQuerySetup';
 
-export class DSLDataSpace_QueryPlugin extends QueryPlugin {
+export class DSLDataSpace_QueryPlugin extends LegendQueryPlugin {
   constructor() {
     super(packageJson.extensions.queryPlugin, packageJson.version);
   }
