+++ packages/legend-studio/src/stores/DSLGenerationSpecification_StudioPlugin_Extension.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { DSL_StudioPlugin_Extension } from './StudioPlugin';
+import type { DSL_StudioPlugin_Extension } from './LegendStudioPlugin';
 
 export interface DSLGenerationSpecification_StudioPlugin_Extension
   extends DSL_StudioPlugin_Extension {
