+++ packages/legend-studio/src/stores/StoreRelational_StudioPlugin_Extension.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { DSL_StudioPlugin_Extension } from './StudioPlugin';
+import type { DSL_StudioPlugin_Extension } from './LegendStudioPlugin';
 import type {
   DatasourceSpecification,
   AuthenticationStrategy,
